pyproject.toml
[project]
name = "aa228v-finalproject"
version = "0.1.0"
description = "Final project repository for AA228V"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "gym>=0.26.2",
    "gymnasium>=1.0.0",
    "ipython>=8.32.0",
    "juliacall>=0.9.24",
    "marimo>=0.11.2",
    "matplotlib>=3.10.0",
    "numpy>=2.2.2",
    "pygame>=2.6.1",
    "pyqt6>=6.8.1",
    "scipy>=1.15.2",
    "torch>=2.5.1",
    "tqdm>=4.67.1",
    "traker>=0.3.2",
]
